Refactor EMSI scraper
